Updated(Feature): Costing Calculations
Methods to get total cost, cost per portion, and cost percent on Recipe Model.

string->float->string shenanigans on portionCostPercentageAsString is because when you call the dividedBy method in getPortionCostPercentage at a scale of 2, it rounds to a whole number. Using a scale of 3 gives the precision and the s->f->s removes the trailing zeros.