New(Feat): Theme Changer
For now, store the theme to the session.
Single Action Controller and Route to change the value in session.
Dropdown menu on main nav
- Split out into backed blade component.
Helper function theme() that returns the asset link for the css based on session variable. Defaults to Lux.
Theme list added to app config.
Added Slate and Sketchy theme.

100% dynamic based on array in config file. To add more to the app, all that needs to be done is put the css file into app/public/css, and add the filename (without .css) to the config array.