Refactor(Design): Form Tags around Livewire Inputs
Instead of triggering the method from a wire:click on a button, add the method to a form wire:submit, and change button to a submit type.
New submit button blade component created under new ls (Livewire Specific) group. Has a `targets` attribute for the livewire method assigned in the form tag to disable and show a spinner while the request is processing.
Refactor all livewire input groups to proper forms.