New(Dev): Ability to disable casting on models.
CanDisableCastingOnModel trait created and added to BaseModel.
Trait creates a $shouldCast boolean static property and static methods to disable and re-enable. A static method updateWithoutCasting that wraps a call to create, disabling casts before and re-enabling after.
DisableCastingInFactory trait created that can be added to a factory that also disables before create/make and re-enables after.
RecipeFactory changed to new concept already for testing.

Main use is to be able to create a model without having to cast the attribute manually first.