Refactor(Model): RecipeItem getCost to action.
Extracted logic from large method getCost on RecipeItem to an Action concept that uses Laravel's pipeline (aliased to ladder) concept. This allows for each step to be extracted into its own class for easier "bite sized" readability. Any future steps to be added can be done by creating a new class and adding it to the steps property on the action.