Refactor(Feat): RecipeItemGetCostAction
Change BothUseEachTypeAsUnit step to BothUnitsAreTheSame.
Add isSameAs method to MeasurementEnum interface and implement. All enums are the same, use trait for method.
Instead of just covering when both the RecipeItem and Ingredient are both using 'each' as the unit, now it will calculate early when both units are the same.