Refactor: Ingredient returns CrossConversion on EachToEach only if values match.
Example: Before, the first CrossConversion model that was of an each to each type would be returned. Now it checks to make sure the values of the EachMeasurement models are correct. So if there is a record that converts Each/Case and one that converts Each/Bunch, the correct one will be used for what the RecipeItem needs.